requirements.txt
flask
google-generativeai
openai
anthropic
perplexipy
